+++ examples/tokio-psk/src/main.rs
@@ -19,16 +19,18 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let config = TlsConfig::new()
         .with_server_name("localhost")
         .with_psk(&[0xaa, 0xbb, 0xcc, 0xdd], &[b"vader"]);
-    let mut rng = OsRng;
-    let mut tls: TlsConnection<FromTokio<TcpStream>, Aes128GcmSha256> = TlsConnection::new(
+    let mut tls = TlsConnection::new(
         FromTokio::new(stream),
         &mut read_record_buffer,
         &mut write_record_buffer,
     );
 
-    tls.open::<OsRng, NoVerify>(TlsContext::new(&config, &mut rng))
-        .await
-        .expect("error establishing TLS connection");
+    tls.open(TlsContext::new(
+        &config,
+        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
+    ))
+    .await
+    .expect("error establishing TLS connection");
 
     tls.write_all(b"ping").await.expect("error writing data");
     tls.flush().await.expect("error flushing data");