+++ src/asynch.rs
@@ -12,7 +12,6 @@ use crate::TlsError;
 use embedded_io::Error as _;
 use embedded_io::ErrorType;
 use embedded_io_async::{BufRead, Read as AsyncRead, Write as AsyncWrite};
-use rand_core::{CryptoRng, RngCore};
 
 pub use crate::config::*;
 #[cfg(feature = "std")]
@@ -71,16 +70,20 @@ where
     ///
     /// Returns an error if the handshake does not proceed. If an error occurs, the connection
     /// instance must be recreated.
-    pub async fn open<'v, RNG, Verifier>(
+    pub async fn open<'v, Provider>(
         &mut self,
-        context: TlsContext<'v, CipherSuite, RNG>,
+        mut context: TlsContext<'v, Provider>,
     ) -> Result<(), TlsError>
     where
-        RNG: CryptoRng + RngCore,
-        Verifier: TlsVerifier<'v, CipherSuite>,
+        Provider: CryptoProvider<CipherSuite = CipherSuite>,
     {
-        let mut handshake: Handshake<CipherSuite, Verifier> =
-            Handshake::new(Verifier::new(context.config.server_name));
+        let mut handshake: Handshake<CipherSuite> = Handshake::new();
+        if let (Ok(verifier), Some(server_name)) = (
+            context.crypto_provider.verifier(),
+            context.config.server_name,
+        ) {
+            verifier.set_hostname_verification(server_name)?;
+        }
         let mut state = State::ClientHello;
 
         while state != State::ApplicationData {
@@ -92,7 +95,7 @@ where
                     &mut self.record_write_buf,
                     &mut self.key_schedule,
                     context.config,
-                    context.rng,
+                    &mut context.crypto_provider,
                 )
                 .await?;
             trace!("State {:?} -> {:?}", state, next_state);