+++ src/handshake/certificate_request.rs
@@ -1,3 +1,4 @@
+use crate::extensions::extension_data::signature_algorithms::SignatureAlgorithms;
 use crate::extensions::messages::CertificateRequestExtension;
 use crate::parse_buffer::ParseBuffer;
 use crate::TlsError;
@@ -7,6 +8,7 @@ use heapless::Vec;
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct CertificateRequestRef<'a> {
     pub(crate) request_context: &'a [u8],
+    pub(crate) extensions: Vec<CertificateRequestExtension<'a>, 6>,
 }
 
 impl<'a> CertificateRequestRef<'a> {
@@ -19,10 +21,11 @@ impl<'a> CertificateRequestRef<'a> {
             .map_err(|_| TlsError::InvalidCertificateRequest)?;
 
         // Validate extensions
-        CertificateRequestExtension::parse_vector::<6>(buf)?;
+        let extensions = CertificateRequestExtension::parse_vector::<6>(buf)?;
 
         Ok(Self {
             request_context: request_context.as_slice(),
+            extensions,
         })
     }
 }
@@ -31,6 +34,7 @@ impl<'a> CertificateRequestRef<'a> {
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct CertificateRequest {
     pub(crate) request_context: Vec<u8, 256>,
+    pub(crate) signature_algorithms: Option<SignatureAlgorithms<16>>,
 }
 
 impl<'a> TryFrom<CertificateRequestRef<'a>> for CertificateRequest {
@@ -43,6 +47,18 @@ impl<'a> TryFrom<CertificateRequestRef<'a>> for CertificateRequest {
                 error!("CertificateRequest: InsufficientSpace");
                 TlsError::InsufficientSpace
             })?;
-        Ok(Self { request_context })
+
+        let mut signature_algorithms = None;
+
+        for ext in cert.extensions {
+            if let CertificateRequestExtension::SignatureAlgorithms(algos) = ext {
+                signature_algorithms = Some(algos)
+            }
+        }
+
+        Ok(Self {
+            request_context,
+            signature_algorithms,
+        })
     }
 }