+++ src/webpki.rs
@@ -4,7 +4,7 @@ use crate::handshake::{
     certificate::{
         Certificate as OwnedCertificate, CertificateEntryRef, CertificateRef as ServerCertificate,
     },
-    certificate_verify::CertificateVerify,
+    certificate_verify::CertificateVerifyRef,
 };
 use crate::TlsError;
 use core::marker::PhantomData;
@@ -89,30 +89,43 @@ static ALL_SIGALGS: &[&webpki::SignatureAlgorithm] = &[
     &webpki::ED25519,
 ];
 
-pub struct CertVerifier<'a, CipherSuite, Clock, const CERT_SIZE: usize>
+pub struct CertVerifier<CipherSuite, Clock, const CERT_SIZE: usize>
 where
     Clock: TlsClock,
     CipherSuite: TlsCipherSuite,
 {
-    host: Option<&'a str>,
+    host: Option<heapless::String<64>>,
     certificate_transcript: Option<CipherSuite::Hash>,
     certificate: Option<OwnedCertificate<CERT_SIZE>>,
     _clock: PhantomData<Clock>,
 }
 
-impl<'a, CipherSuite, Clock, const CERT_SIZE: usize> TlsVerifier<'a, CipherSuite>
-    for CertVerifier<'a, CipherSuite, Clock, CERT_SIZE>
+impl<CipherSuite, Clock, const CERT_SIZE: usize> CertVerifier<CipherSuite, Clock, CERT_SIZE>
 where
-    CipherSuite: TlsCipherSuite,
     Clock: TlsClock,
+    CipherSuite: TlsCipherSuite,
 {
-    fn new(host: Option<&'a str>) -> Self {
+    pub fn new() -> Self {
         Self {
-            host,
+            host: None,
             certificate_transcript: None,
             certificate: None,
             _clock: PhantomData,
         }
+    }
+}
+
+impl<CipherSuite, Clock, const CERT_SIZE: usize> TlsVerifier<CipherSuite>
+    for CertVerifier<CipherSuite, Clock, CERT_SIZE>
+where
+    CipherSuite: TlsCipherSuite,
+    Clock: TlsClock,
+{
+    fn set_hostname_verification(&mut self, hostname: &str) -> Result<(), TlsError> {
+        self.host.replace(
+            heapless::String::try_from(hostname).map_err(|_| TlsError::InsufficientSpace)?,
+        );
+        Ok(())
     }
 
     fn verify_certificate(
@@ -121,13 +134,13 @@ where
         ca: &Option<Certificate>,
         cert: ServerCertificate,
     ) -> Result<(), TlsError> {
-        verify_certificate(self.host.clone(), ca, &cert, Clock::now())?;
+        verify_certificate(self.host.as_deref(), ca, &cert, Clock::now())?;
         self.certificate.replace(cert.try_into()?);
         self.certificate_transcript.replace(transcript.clone());
         Ok(())
     }
 
-    fn verify_signature(&mut self, verify: CertificateVerify) -> Result<(), TlsError> {
+    fn verify_signature(&mut self, verify: CertificateVerifyRef) -> Result<(), TlsError> {
         let handshake_hash = unwrap!(self.certificate_transcript.take());
         let ctx_str = b"TLS 1.3, server CertificateVerify\x00";
         let mut msg: Vec<u8, 130> = Vec::new();
@@ -146,7 +159,7 @@ where
 fn verify_signature(
     message: &[u8],
     certificate: ServerCertificate,
-    verify: CertificateVerify,
+    verify: CertificateVerifyRef,
 ) -> Result<(), TlsError> {
     let mut verified = false;
     if !certificate.entries.is_empty() {