+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Add missing implementation to support Client Certificate Authorization (#135)
+
 ## 0.17.0 - 2024-01-06
 
 - Update to stable rust